[REFACTOR] Moves auth controller functions
